+++ src/main/java/domain/generator/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.generator;
 
 public interface BooleanGenerator {
