+++ src/main/java/domain/generator/RandomBooleanGenerator.java
@@ -1,10 +1,11 @@
-package domain;
+package domain.generator;
 
 import java.util.Random;
 
 public class RandomBooleanGenerator implements BooleanGenerator {
 
     private final Random random = new Random();
+
     @Override
     public boolean generate() {
         return random.nextBoolean();
