+++ src/main/java/domain/ladder/PointJudge.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.List;
