+++ src/main/java/domain/player/Player.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import java.util.Objects;
 
@@ -34,8 +34,12 @@ public class Player {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
         Player player = (Player) o;
         return Objects.equals(name, player.name);
     }
