+++ src/test/java/domain/ladder/LineTest.java
@@ -1,7 +1,7 @@
-package domain;
+package domain.ladder;
 
+import domain.generator.BooleanGenerator;
 import org.junit.jupiter.api.DisplayName;
-
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -10,7 +10,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.List;
 import java.util.stream.Stream;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class LineTest {
