+++ src/test/java/domain/ladder/PointTest.java
@@ -1,9 +1,9 @@
-package domain;
+package domain.ladder;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 
 class PointTest {
